test: Add support for dbus and tls IPC frontend.
For the test harness to launch test executables against the different
IPC frontends we've added the TABRMD_TEST_TCTI environment variable.
This variable is used to pass for the name for the TCTI library to
the test harness which will instante the TCTI for the test being
executed.

It's not possible to dynamically load either of the two local TCTIs
(w/o doing stuff with the linker that we probably shouldn't) so we
check for their names manually in this variable and link to their
init functions directly. Any other string is treated like an unknown
TCTI and the dynamic loading API is used. This is intended to allow
execution of test cases directly against the mssim or whatever other
TCTI you may have handy.

The int-test-setup.sh is also updated to take a new parameter:
--ipc-mode that will accept the string "dbus" or "tls". This configures
the tpm2-abrmd instance to use the specified ipc-frontend. The
environment for the test executable is configured using the two new
environment variables from above to use the appropriate TCTI and
configuration.

Tests can be run in parallel so long as they're run against the mssim.
All tests executed using the device TCTI must be run serially since
only one connection can be made at a time.

Signed-off-by: Philip Tricca <flihp@twobit.us>