build: Replace hwtpm config option with test-tcti and test-ipcfront.
The 'hwtpm' configure option would cause the test harness to use the
device TCTI instead of the default mssim. This commit replaces this
behavior allowing the caller to select any TCTI, device, mssim or
otherwise.

This commit also adds an option to allow the caller to select which of
the supported IPC frontends that will be used by the integration test
harness.

Both of these configure options can be overriden by assigning the
appropriate variable in the makefile when executing the make command.

Signed-off-by: Philip Tricca <flihp@twobit.us>