build: Add configure option to enable SELinux policy build.
This commit adds the `--with-refpol` option to the configure script. If
provided (disabled by default) the SELinux policy module is built and
installed by the default make target. The existing functionality
provided by the hand-rolled Makefile in the $(srcdir)/selinux directory
remains unchanged to preserve backward compatibility but should removed
as much as possible in the future.

Some weirdness in this commit is required due to weirdness in the
SELinux build infrastructure. Specifically we cannot simply include the
SELinux devel Makefile in Makefile.am. Also we cannot invoke the SELinux
devel Makefile from the Makefile generated by autotools in $(srcdir) as
it requires the SELinux policy sources be in the directory where make
was invoked. This forces us to invoke `make -C selinux` from the
Makefile in $(srcdir) to kick off the Makefile under $(srcdir)/selinux.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>